+++ Tools/CommonByCode/Http/Http/Heads/HtpUserAgents.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Jxl.Http
+namespace Common.HttpHead
 {
     internal static class HttpUserAgents
     {
