+++ Tools/CommonByCode/Http/Http/Heads/HttpContentTypes.cs
@@ -1,4 +1,4 @@
-﻿namespace Jxl.Http
+﻿namespace Common.HttpHead
 {
     /// <summary>
     /// Content-Type首部
