+++ Tools/CommonByCode/Http/Http/Loggers/IHttpLogger.cs
@@ -1,10 +1,10 @@
-﻿using Jxl.Http;
+﻿using Common.HttpException;
 using Monitor.Common;
 using System;
 using System.IO;
 using System.Net.Http;
 
-namespace Jxl.Logging
+namespace Common.Logging
 {
     /// <summary>
     /// HTTP请求的日志记录器，记录成功的<see cref="HttpRequestMessage"/>与<see cref="HttpResponseMessage"/>信息。
@@ -38,64 +38,5 @@ namespace Jxl.Logging
         public string DataType { get; internal set; }
 
         internal HttpMessage() { }
-    }
-
-
-    /// <summary>
-    /// 默认日志记录者
-    /// 日志保存在"程序根目录\GlobalLogs\yyyy-MM-dd\GUID\"目录下
-    /// </summary>
-    internal class DefaultLogger : ILogger
-    {
-        public string _path;
-        public DefaultLogger(string path = "")
-        {
-            _path = path;
-        }
-
-        public virtual void Log(string message)
-        {
-            if (string.IsNullOrWhiteSpace(_path))
-            {
-                _path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs", DateTime.Now.ToString("yyyy-MM-dd"), Guid.NewGuid().ToString("N"));
-            }
-
-            if (!Directory.Exists(_path))
-            {
-                Directory.CreateDirectory(_path);
-            }
-
-            File.AppendAllText(Path.Combine(_path, "log.log"), $"{DateTime.Now.ToString("hh:mm:ss")}>>>{message}");
-        }
-    }
-
-    /// <summary>
-    /// 默认HTTP日志记录者
-    /// 日志保存在"程序根目录\GlobalLogs\yyyy-MM-dd\GUID\"目录下
-    /// </summary>
-    internal class DefaultHttpLogger : IHttpLogger
-    {
-        public string _path;
-        public DefaultHttpLogger(string path = "")
-        {
-            _path = path;
-        }
-
-        public void Log(HttpMessage httpMessage)
-        {
-            if (string.IsNullOrWhiteSpace(_path))
-            {
-                _path = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Logs", DateTime.Now.ToString("yyyy-MM-dd"), Guid.NewGuid().ToString("N"));
-            }
-
-            if (!Directory.Exists(_path))
-            {
-                Directory.CreateDirectory(_path);
-            }
-
-            var path = string.IsNullOrWhiteSpace(httpMessage.DataType) ? Path.Combine(_path, $"{Tool.GetUnixTimestamp()}.html")
-                : Path.Combine(_path, $"{Tool.GetUnixTimestamp()}_{httpMessage.DataType}.html");
-            File.AppendAllText(path, httpMessage.LogContent);
-        }
     }
 }
