+++ Tools/CommonByFramework/MailNotice/MessageFormat.cs
@@ -6,7 +6,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
-namespace Message
+namespace Common.Message
 {
     /// <summary>
     /// 消息格式化
